Update README.md
I learned python for the first time and did my homework.
Unfortunately, I'm not good at English. Please refer to this point.

Here are the rules:
1) Your rocket starts at position 100m above the earth. Initial velocity is 0m/s. 
2) Gravity pulls you down at acceleration 10m/s^2
3) You have 100 liters of fuel
4) You have thrusters that can be powered from 0 to 20. 1 thrust = 1 liter of fuel
5) Total acceleration = gravity + thrusters (minimum -10, maximum +10)

I did my homework. but This is the first time that I learned python. I want to use function(def) but i can't use it well.